+++ engine/flink/tests/src/test/scala/pl/touk/nussknacker/engine/flink/minicluster/scenariotesting/FlinkMiniClusterScenarioTestRunnerSpec.scala
@@ -162,8 +162,8 @@ class FlinkMiniClusterScenarioTestRunnerSpec
       invocationResults("out") shouldBe
         List(ExpressionInvocationResult(s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-1", "Value", variable(11)))
 
-      results.externalInvocationResults("proc2") shouldBe List(
-        ExternalInvocationResult(
+      results.externalInvocationResults("proc2").map(r => (r.contextId, r.name, r.value)) shouldBe List(
+        (
           s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-1",
           "logService",
           variable("0-collectedDuringServiceInvocation")
@@ -688,10 +688,10 @@ class FlinkMiniClusterScenarioTestRunnerSpec
           .futureValue
 
       results.invocationResults("proc2").map(_.contextId) should contain only (
-        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-1-end1",
-        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-2-end1",
-        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-0-end2",
-        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-2-end2"
+        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-1-left-end1",
+        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-2-left-end1",
+        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-0-right-end2",
+        s"$scenarioName-$sourceNodeId-$firstSubtaskIndex-2-right-end2"
       )
 
       results